PCL build is now only a build configuration (maintaining .NET 3.5, .NET 4.0 and .NET 4.5 builds)
