+++ src/Colourful/Colors/RGBColor.cs
@@ -25,6 +25,10 @@ using Vector = System.Collections.Generic.IReadOnlyList<double>;
 using Matrix = System.Collections.Generic.IReadOnlyList<System.Collections.Generic.IReadOnlyList<double>>;
 #endif
 
+#if (!PCL)
+using System.Drawing;
+#endif
+
 namespace Colourful
 {
     /// <summary>
@@ -80,6 +84,23 @@ namespace Colourful
         {
             WorkingSpace = workingSpace;
         }
+
+#if (!PCL)
+
+        /// <remarks>Uses <see cref="DefaultWorkingSpace"/> as working space.</remarks>
+        public RGBColor(Color color)
+            : this(color, DefaultWorkingSpace)
+        {
+        }
+
+        /// <param name="workingSpace"><see cref="RGBWorkingSpaces"/></param>
+        public RGBColor(Color color, IRGBWorkingSpace workingSpace)
+            : base(((double)color.R) / 255, ((double)color.G) / 255, ((double)color.B) / 255)
+        {
+            WorkingSpace = workingSpace;
+        }
+
+#endif
 
         #endregion
 
@@ -173,6 +194,37 @@ namespace Colourful
         }
 
         #endregion
+
+#if (!PCL)
+
+        #region Color conversions
+
+        public Color ToColor()
+        {
+            return this;
+        }
+
+        public static implicit operator Color(RGBColor input)
+        {
+            if (input == null)
+                return new Color();
+
+            var r = (byte)Math.Round(input.R * 255).CropRange(0, 255);
+            var g = (byte)Math.Round(input.G * 255).CropRange(0, 255);
+            var b = (byte)Math.Round(input.B * 255).CropRange(0, 255);
+            var output = Color.FromArgb(r, g, b);
+            return output;
+        }
+
+        public static explicit operator RGBColor(Color color)
+        {
+            var output = new RGBColor(color);
+            return output;
+        }
+
+        #endregion
+
+#endif
         
         #region Overrides
