--- src/Colourful/Implementation/Utils/PublicExtensions.cs
@@ -1,17 +1,0 @@
-﻿
-namespace Colourful.Implementation
-{
-    public static class PublicExtensions
-    {
-        public static double CropRange(this double value, double min, double max)
-        {
-            if (value < min)
-                return min;
-
-            if (value > max)
-                return max;
-
-            return value;
-        }
-    }
-}
